relocate the benchmarks to follow Rust best practices
