--- oxen-rust/crates/benches/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "oxen-bench"
-version = { workspace = true }
-edition = { workspace = true }
-publish = false
-
-[dependencies]
-criterion = { workspace = true }
-liboxen = { path = "../lib" }
-rand = { workspace = true }
-tokio = { workspace = true }
-uuid = { workspace = true }
-
-[[bench]]
-name = "benchmark"
-path = "src/benchmark.rs"
-harness = false
